organisational
removed unnecessary file, organised functions into clearer groupings and into a file with a clearer name
