+++ src/__init__.py
@@ -1,7 +1,7 @@
 # this is the namespace thing, you need to have the filepath there
 # Because by the nature of psychological experiments, the GUI is a required part of using this
 # this __init__ file may end up only containing a function to launch the GUI
-from src.GUItesting import InitGUI
+from src.functions import InitGUI
 
 # The only other thing it may contain is a function that takes a config and starts an experiment from there
